Add read/write operations
Writing should be done always with offset 0, or -EINVAL will
be returned. This means, writting should be atomic.

If the write count is > than capacity, a first write will fill
the whole capacity and the second write with bigger offset will
fail, leaving uncomplete data in the content. Should it return
an error if write size is > than capacity instead?