Add suport for polling by multiples proccess
Add per open info about if the device has been read since the last
write and check this for the polling.

Open and release functions are needed.

As write has to access all per-open info, every per-open info is also
included as a linked list in the global device info.